outage-processor/requirements.txt
flask
google-cloud-pubsub
google-cloud-firestore